cleanup: Make no_free_ptr() __must_check
recent discussion brought about the realization that it makes sense for
no_free_ptr() to have __must_check semantics in order to avoid leaking
the resource.

Additionally, add a few comments to clarify why/how things work.

All credit to Linus on how to combine __must_check and the
stmt-expression.

[ Upstream commit 85be6d842447067ce76047a14d4258c96fd33b7b ]
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Link: https://lkml.kernel.org/r/20230816103102.GF980931@hirez.programming.kicks-ass.net
Backported-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I6d6cf53857d144343a031c69a629655bb7e8e54a