KVM: arm64: Move the clean of dcache to the map handler
We currently uniformly clean dcache in user_mem_abort() before calling
the fault handlers, if we take a translation fault and the pfn is
cacheable. But if there are concurrent translation faults on the same
page or block, clean of dcache for the first time is necessary while
the others are not.

By moving clean of dcache to the map handler, we can easily identify
the conditions where CMOs are really needed and avoid the unnecessary
ones. As it's a time consuming process to perform CMOs especially when
flushing a block range, so this solution reduces much load of kvm and
improve the efficiency of creating mappings.

Mainline: Open-Source
Signed-off-by: Zou Wei <zouwei@phytium.com.cn>
Signed-off-by: Li Mingzhe <limingzhe1839@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I7dbad984563c955faf7531eecea94b99880e73c6