bpf: Add bpf_cpumask_weight() kfunc
It can be useful to query how many bits are set in a cpumask. For
example, if you want to perform special logic for the last remaining
core that's set in a mask. Let's therefore add a new
bpf_cpumask_weight() kfunc which checks how many bits are set in a mask.

[ Upstream commit a6de18f310a511278c1ff16b96eb2d500eada725 ]
Signed-off-by: David Vernet <void@manifault.com>
Acked-by: Yonghong Song <yonghong.song@linux.dev>
Link: https://lore.kernel.org/r/20231207210843.168466-2-void@manifault.com
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Backported-by: Huang Shaobo <huangshaobo2075@phytium.com.cn>
Change-Id: Ib97eb1bff51b0eec70f5eaab3781a1eccbb4a585