KVM:arm64: Fixed gpu passthrough to vm on arm64
The vfio memory type is set to DEVICE_nGnRnE, and the KVM
memory type is set to DEVICE_nGnRnE, but in guestOS, the
memory type is set to Device_ngnrne. The memory attribute
of all I/O memory mappings is set to MT_NORMAL_NC, and the
graphics card is no exception as a peripheral in guestOS.
DEVICE_nGnRnE is the strongest according to the memory
attribute rule, which requires that the memory addresses
of devices in VMS must be aligned. Therefore, the error
reported when the graphics card is transparent transmission
is also the memory address alignment error.

In this patch, if the passthrough device is a graphics card
device, the memory attribute is changed to MT_NORMAL_NC
instead of DEVICE_nGnRnE, so that memory alignment errors
can be avoided.

Mainline: Open-Source
Signed-off-by: Peng Mengguang<pengmengguang@phytium.com.cn>
Signed-off-by: Li Mingzhe <limingzhe1839@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I28d53367317a28f46d51a501603932cc0fd769f6