hda: phytium: Re-write function for dma_ops alloc
This patch aids to fix the problem when "dma-coherent" is described
in DT on the board pd1904 and pd2008. The board pd2308 does not have
the same problem. The old patch used api for ACPI when match dts node
which may not meet the specification.

This patch takes hda node out of soc node to avoid the influence
of "dma-coherent".

Signed-off-by: Dai Jingtao <daijingtao1503@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I41069ea7dd855a491de14949d301f13f5bc567f7