drm/ttm: Fix qemu bad address error with phytium vgpu acceleration enabled
With phytium virtual gpu acceleration method in qemu windows guest OS,
guest OS driver needs to share GPU device memory with host and requires
the shared buffer mapping remains valid, but the ttm map implemenataion
did not account for this usecase. So we add a ref count for every pages
of the shared buffer.

Signed-off-by: Liu Tao <liutao@phytium.com.cn>
Signed-off-by: Li Mingzhe <limingzhe1839@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I20cf405757e6aa7fd6953fb603dc8c455badefde