can:phytium: Fix the bug of runtime PM when removing module.
This patch fixes the bug of runtime PM operation when removing module.
When removing the can module,the private management structure of can
has been freed,but at this time the runtime PM is not disabled,so
the runtime PM operation is called again.The system encountered an
exception due to accessing data from a dangling pointer.

Signed-off-by: Wu Jinyong <wujinyong1788@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I86eea10a38291c8fbde771c3cd92577b90d18a0a