sched_ext: s/SCX_RQ_BALANCING/SCX_RQ_IN_BALANCE/ and add SCX_RQ_IN_WAKEUP
SCX_RQ_BALANCING is used to mark that the rq is currently in balance().
Rename it to SCX_RQ_IN_BALANCE and add SCX_RQ_IN_WAKEUP which marks whether
the rq is currently enqueueing for a wakeup. This will be used to implement
direct dispatching to local DSQ of another CPU.

[ Upstream commit f47a818950dd5e5d16eb6e9c1713bb0bc61649cd ]
Signed-off-by: Tejun Heo <tj@kernel.org>
Acked-by: David Vernet <void@manifault.com>
Backported-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: Ie404ad2cb30dbb6a6c1ad9ced89b18964c95e23b