i2c: phytium: Fix stop interrupt in slave mode
This patch is to fix an issue where the STOP interrupt
signal is triggered too early in slave mode.

The complete data processing flow is that all other
interrupt signals are processed,and the STOP interrupt
signal is processed last.However,the STOP interrupt
signal is triggered too early,which terminates the later
data processing flow premanturely.

Modify the interrupt handler,when there is a data
interrupt,empty the RXFIFO to make sure all the data
are processed,and then process the STOP interrupt.

Mainline: Open-Source
Signed-off-by: Wu Jinyong <wujinyong1788@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I41296a9aad4a33e14c17d4e906bfac68a94a559a