arm64: add support for machine check error safe
During the processing of arm64 kernel hardware memory
errors(do_sea()), if the errors is consumed in the kernel,
the current processing is panic. However, it is not optimal.

Take uaccess for example, if the uaccess operation fails due
to memory error, only the user process will be affected, kill
the user process and isolate the user page with hardware
memory errors is a better choice.

This patch only enable machine error check framework, it add
exception fixup before kernel panic in do_sea() and only
limit the consumption of hardware memory errors in kernel
mode triggered by user mode processes. If fixup successful,
panic can be avoided.

Signed-off-by: Tong Tiangen <tongtiangen@huawei.com>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: Ic6097eff14e6e6b86318b5e6f495373e264bb67d