scx_flatcg: Use a user DSQ for fallback instead of SCX_DSQ_GLOBAL
scx_flatcg was using SCX_DSQ_GLOBAL for fallback handling. However, it is
assuming that SCX_DSQ_GLOBAL isn't automatically consumed, which was true a
while ago but is no longer the case. Also, there are further changes planned
for SCX_DSQ_GLOBAL which will disallow explicit consumption from it. Switch
to a user DSQ for fallback.

[ Upstream commit c9c809f4137c3c0f962226c10d245d3ce2fd5b7c ]
Signed-off-by: Tejun Heo <tj@kernel.org>
Acked-by: David Vernet <void@manifault.com>
Backported-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: Ia20634abbe4d821901350ced0dc894272e5113b6