thermal: boost: phytium: Automatic enable/disable of BOOST feature
This patch provides auto disable/enable operation for boost. When
the device temperature rises above 75% of the smallest trip point,
the boost is disabled.
In that moment thermal monitor workqueue will monitors if the device
cools down. When the device temperature drops below 75% of the
smallest trip point, the boost is enabled again.

Mainline: NA
Signed-off-by: Li Jiayi <lijiayi1493@phytium.com.cn>
Signed-off-by: Li Mingzhe <limingzhe1839@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I52ec7cc4b2d3daad35c1835909e6be3cb8f1ae22