net/macb: Change the minimum rx ring size to 128
If the budget is equal to 64, gem_rx can exit after
receiving a maximum of budget packets.
If the ring size is changed to 64, a descriptor that
has not been refilled is considered to be a valid
descriptor to be received, and the error
"inconsistent Rx descriptor chain" is reported.
So we need to limit the minimum rx ring size to be
greater than 64.

Mainline: Open-Source
Signed-off-by: Li Wencheng <liwencheng@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I3607a3b229bf88d02c2a78e4bb2885e678a29279