net/macb: Fixed the CPU stall issue in S4 test
In the S4 test, it is possible that macb_mac_link_down concurrency with
macb_tx_interrupt caused an endless loop, which in turn caused
the CPU to stall.
So, it is necessary to clear the tx_used bit in macb_mac_link_down
to break the condition for the macb_tx_interrupt to enter the loop.

Mainline: Open-Source
Signed-off-by: Li Wencheng <liwencheng@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I2186315cb6fa7c383923a30ded061a1d0a6eb9a1