arm64: Add support for machine check error safe
During the processing of arm64 kernel hardware memory
errors(do_sea()), if the errors is consumed in the kernel,
the current processing is panic. However, it is not optimal.

Take uaccess for example, if the uaccess operation fails
due to memory error, only the user process will be affected,
kill the user process and isolate the user page with
hardware memory errors is a better choice.

This patch only enable machine error check framework, it
add exception fixup before kernel panic in do_sea() and
only limit the consumption of hardware memory errors in
kernel mode triggered by user mode processes. If fixup
successful, panic can be avoided.

Mainline: NA
Signed-off-by: Tong Tiangen <tongtiangen@huawei.com>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: Ibd174a01a24e21f080d791f73ec7d71969be24bf