sched_ext: Add option -l in selftest runner to list all available tests
The selftest runner currently allows selecting tests via the -t
option. This patch adds a new -l option that lists all available tests,
providing users with an overview of the tests they can choose from. This
enhancement is especially useful for scripting and automation purposes,
making it easier to discover and run tests.

[ Upstream commit a73bca3d9cc0a27dd7061a55841a1d752ec7365a ]
Signed-off-by: Shizhao Chen <shichen@redhat.com>
Signed-off-by: Tejun Heo <tj@kernel.org>
Backported-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I52390be82512f90c22c1e6b5cf8ec718123266c4