i2c-2.0: phytium: Fixed the bug of combining two interrupts.
When an I2C interrupt is triggered,the CPU is processing
other interrupts at this time,or there are more interrupts
queued in front of the I2C interrupt,resulting in this current
I2C interrupt not being able to respond in time.Then another
I2C interrupt comes,so that by the time the I2C interrupt is
responded to,the interrupt handler can only be triggered
once,and only one transaction is handled,thus leading to
incomplete transaction processing.

By using interrupt plus polling,when an interrupt is triggered,
all pending transactions are polled.

Mainline: NA
Signed-off-by: Wu Jinyong <wujinyong1788@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I64be2029c82ff134b9b129749b67f24e0f7b404f