codec-v2: phytium: Bugfix the NULL pointer issue in S3
When entering suspend mode, the controller mistakenly releases the
memory space for codec registers due to an incorrect if statement.
Upon resuming, it releases the same memory space again. which will
cause the NULL pointer issue.

So modify if statement in suspend, and add a new function to print
codec registers.

Mainline: NA
Signed-off-by: Zhou Zheng <zhouzheng2069@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: Ib002be01922c2e6e2dbde67eedd42fa76a2aa623