mm/init: Add KASAN quirk for phytium ps23064 SoC
Enabling KASAN in the Phytium ps23064 SoC will result in a reduction
in linear mapping space, leading to a failure and panic in allocating
percpu variables. This patch avoid this problem throuth removing high
physical addresses, but it should be note that device should not be
attach to die 5-7.

Mainline: NA
Signed-off-by: Huang Shaobo <huangshaobo2075@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I3fbd33044d97c1739ac1b4889b79d39435e33d6a