devfreq: Phytium: Update some new functions for DMU driver
This patch modifies and adds the following functions:
1). On account of DMU and DDR PMU drivers operate PMU
registers at the same time, which will result in conflict.
So the register operation of se in dmufreq is transferred
to the upper driver.

2). The notification chain of dmufreq to DDR PMU is added
in order to suspend dmufreq's register action and maintain
the rate at the current frequency when the PMU driver is loaded.

3). Add suspend and resume features.

Mainline: Open-Source
Signed-off-by: Li Mingzhe <limingzhe1839@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: Iaee1415a791fe2e0fee7bcbd8e5c4730e5aae27e