pswiotlb: Check if min_align_mask of dma is valid
The min_align_mask of dma could be invalid or uninitialized in
some kernel versions(e.g. v4.19). Add the processing procedure
to determine whether the value is valid or not. The min_align_mask
can not be used to calculate the offset address when it is invalid.

MAINLINE: Open-Source
Signed-off-by: Cui Chao <cuichao1753@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: Ifb738a3a9eecee87a4f4dccbb96b7aa992529c39