i2s-v2: phytium: Avoid using mutex in i2s interrupt
In interrupt context, funcitons that may cause blocking are not
allowed. However, in the i2s interrupt handler, the function
snd_pcm_period_elapsed attempted to use a mutex due to nonatomic,
which will result in a call trace error. To resolve this, another
thread is used to execute the function.

Mainline: Open-Source
Signed-off-by: Zhou Zheng <zhouzheng2069@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I0828ca2278e8fbaf927e1e5097500c3d90a15735