i2s-v2: phytium: Remove nonatomic attribute
In interrupt context, funcitons that may case blocking are not allowed.
However, in the i2s interrupt handler, the function
snd_pcm_period_elapsed attempted to use a mutex due to nonatomic,
resulting in a call trace error.

To resolve this, the nonatomic attribute and the mutex of sharemem are
removed.sharemem is divied into I2S and I2S_GPIO two parts, and trigger
command was replaced by hw_reg.

Mainline: Open-Source
Signed-off-by: Zhou Zheng <zhouzheng2069@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: Ie3e1415cd29513ab5cc8df53a2d1ec1943d25aa3