i2c-2.0: phytium: Fixed the bug of delayed response to interrupts
Before the modification,when the AP core was responding to other
interrupt handlers,it could not respond in time to the I2C
interrupt cast by the RV core,and at this time,if the RV core
finished processing all the packets and judged that it was not
the last packet of data,it would cast an error and end the
transmission.
After the modification,the RV core no longer considers this
condition as an error,but switches to other transactions for
processing,and when the AP core's interrupt program gets a
response,it triggers the RV core to start the I2C transaction
transmission again.

Mainline: Open-Source
Signed-off-by: Wu Jinyong <wujinyong1788@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I4eb257387317325f4176c7c1c6e8e84b320cd70b