i2c-2.0: phytium: Fixed the bug that processed unupdated data
In the RV interrupt handler, the head pointer is updated, then an AP
interrupt is triggered. In the AP interrupt handler, the RV interrupt
is cleared, then the head pointer is read. If the head pointer does
not match the tail pointer managed by AP, AP continues to evaluate the
shared data content. When the AP responds to the last interrupt
triggered by RV, if RV has just updated the head pointer, the AP reads
the head pointer value for this interrupt and performs the
corresponding data update. After the AP exits this interrupt, it is
triggered again by a new interrupt from RV. However, the AP has already
completed processing in the previous interrupt, and RV has not yet
updated the data. The AP determines that the data state has not been
updated and reports an error.

After modification, the AP's interrupt handler checks the head pointer.
If the head pointer values differ between consecutive interrupts, it
indicates RV has updated the data, and the AP proceeds with processing.

Mainline: Open-Source
Signed-off-by: Wu Jinyong <wujinyong1788@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I5622b5b8187c5448f8b5f1ae066ec40ffad3fba6