usb: xhci: Add the method to reactivate the USB controller
Sometimes, when the USB controller experiences a timeout for the stop
EP endpoint command for a timeout for processing commands, xhci driver
will shutdown the USB controller. If you want to reactivate the USB
controller, you need to follow the reactivation operation process.
note: This patch is only applicable to the Phytium platform.

Mainline: Open-Source
Signed-off-by: Chen Zhenhua <chenzhenhua@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I1444da9ccb1a3aabaa891aff66025cc6c494aa3f