i2s: phytium: Add use_dai_pcm_id attribute
Pulseaudio will read pcm device id and find the corresponding
JACK based on this id. It determines whether to generate a
pulseaudio device. i.e. whether the corresponding audio device
appears in the system based on the plug status of the JACK.

Add use_dai_pcm_id attribute to specify device id, so that
even when devices are loaded out of order, the correct id can
still be obtained for the device. This ensures correct
pulseaudio behavior.

Mainline: Open-Source
Signed-off-by: Dai Jingtao <daijingtao1503@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I3b836ca7e1f5963547e145f1a7606a041f3adc52