i2s-v2: phytium: Defer I2S probe procedure
The problem is that when the hardware card is not inserted, it
causes sound card loading to fail due to undefined behavior from
headphone detection. This detection is in the I2S driver's probe
function, but I2S cannot detect whether a daughter card actually
exists. Therefore, the codec's probe should execute first and return
directly if not daughter card is found.

Mainline: Open-Source
Signed-off-by: Dai Jingtao <daijingtao1503@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Change-Id: I43d2ec94560e1a98275f35ce6912a6eae5721aa9